requirements.txt
accelerate==0.34.2
av==14.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
diffusers==0.32.2
einops==0.8.1
filelock==3.17.0
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
psutil==6.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sentencepiece==0.2.0
sympy==1.13.3
tokenizers==0.20.3
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.1
transformers==4.46.1
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.3.0
zipp==3.21.0
